+++ app/templates/app/team_page.html
@@ -30,16 +30,20 @@ base.html and static are in development
       </div>
     </div>
     {% comment %} the teams context will be sorted from the views {% endcomment %}
-        {% for player in teammates %}
-            <div class="bs-component">  
-                <div  class="row">
-                    <div class="col-1">
-                        <h3> {{ player.username }} </h3>
-                        <h3> {{ player.team.name }} </h3>
-                    </div>
-                </div>
-            </div>    
-        {% endfor %}
+    {% for team_name in team %}
+    <div class="bs-component">  
+        <div  class="row">
+            <div class="col-160">
+              <h4>{{team_name.name}} - {{team_name.id}}</h4>
+              {%for player_name in teammates%}
+                {% if player_name.team.name == team_name.name %}
+                  <p>{{ player_name.username }} - {{ player_name.id }}</p>
+                {% endif %}
+              {%endfor%}
+            </div>
+        </div>
+    </div>    
+    {% endfor %}
 
         <div style='padding:10px ; text-align:center'>
             {% comment %} <button href="{% url 'app:start_page' %}" >Leave Room</button> {% endcomment %}
